requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
drf-yasg==1.21.5
coreschema==0.0.4
cron-descriptor==1.4.0
Django==4.2.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-timezone-field==5.1
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.3
prompt-toolkit==3.0.39
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
psycopg2-binary>=2.9
gunicorn>=20.0.0
djangorestframework-simplejwt>=5.2.2